Add hiccup ns
Because this behaves different than the other renderers (returns a
hiccup data structure rather than a collection of strings to print),
it is not included in the list of built in renders in the core
namespace. Instead, one can use `crockery.hiccup/table` directly, with
an almost identical API as `crockery.core/table`.