Add :render-cell and :render-title options
Like the docstring says, these provide a way to custom render the data
returned by `:key-fn`. While the defaults are useful for human
consumption and reading, there are times when raw, readable Clojure
data is useful to see inside of a tabular format. In those cases,
using `pr-str` is a great way to handle it.