Allow table-wide column default options with `:defaults`
Sometimes, every column needs to be printed with `pr-str`, almost
every column needs to align right, or every column title needs to
render in a non-default way. For all of these cases, specifying the
same option in a colspec over an over is tedious, and can easily
merged in using the new `:defaults` option.