Add new table option `:max-width`, default to terminal width
This adds a new table option, `:max-width`, which resizes columns when
they would exceed the specified width. This only affects fixed-width
renderers (:org, :gfm, :fancy, :tsv).

In clj mode, this will attempt to detect the size of your tty device.
Because this is not 100% portable, although widely available, when the
max width will not be set if it can't be detected and isn't passed in.
This may change in the future, as having a reasonable default (200?)
would probably be better than unbounded output.